+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ### Features
+* Core | Define `clientOrderId` in Orders API [#107](https://github.com/Reiryoku-Technologies/Mida/pull/107) | [#102](https://github.com/Reiryoku-Technologies/Mida/issues/102)
+* Binance | Integrate `clientOrderId` [#107](https://github.com/Reiryoku-Technologies/Mida/pull/107) | [#102](https://github.com/Reiryoku-Technologies/Mida/issues/102)
+* Bybit | Integrate `clientOrderId` [#107](https://github.com/Reiryoku-Technologies/Mida/pull/107) | [#102](https://github.com/Reiryoku-Technologies/Mida/issues/102)
+* cTrader | Integrate `clientOrderId` [#107](https://github.com/Reiryoku-Technologies/Mida/pull/107) | [#102](https://github.com/Reiryoku-Technologies/Mida/issues/102)
 * Core | Enforce all events order from most specific to most generic (m15PeriodUpdate -> periodUpdate -> update) [#104](https://github.com/Reiryoku-Technologies/Mida/pull/104)
 * cTrader | Add `requestedProtection` to orders [#103](https://github.com/Reiryoku-Technologies/Mida/pull/103)
 * Core | Add `requestedProtection` to `MidaOrder` for accessing an order's stop loss, take profit and trailing stop loss [#103](https://github.com/Reiryoku-Technologies/Mida/pull/103)